make controller do the transfer between entities and dtos
