+++ src/main/java/com/chatop/rental/controllers/RentalController.java
@@ -6,6 +6,7 @@ import com.chatop.rental.dto.ErrorResponseDTO;
 import com.chatop.rental.dto.RentalDTO;
 import com.chatop.rental.dto.RentalListDTO;
 import com.chatop.rental.dto.RentalListResponseDTO;
+import com.chatop.rental.entities.Rental;
 import com.chatop.rental.dto.MsgResponseDTO;
 import com.chatop.rental.services.RentalService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -15,6 +16,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -24,6 +27,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/rentals")
@@ -45,7 +49,9 @@ public class RentalController {
       return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
     }
 
-    List<RentalDTO> rentals = rentalService.getAllRentals();
+    List<RentalDTO> rentals = rentalService.getAllRentals().stream()
+        .map(this::convertToListDTO)
+        .collect(Collectors.toList());
     RentalListResponseDTO response = new RentalListResponseDTO(rentals);
     return ResponseEntity.ok(response);
   }
@@ -58,7 +64,7 @@ public class RentalController {
   })
   @GetMapping("/{id}")
   public ResponseEntity<RentalDTO> getRentalById(@PathVariable Long id) {
-    RentalDTO rental = rentalService.getRentalById(id);
+    RentalDTO rental = convertToDTO(rentalService.getRentalById(id));
     return ResponseEntity.ok(rental);
   }
 
@@ -100,5 +106,37 @@ public class RentalController {
     rentalService.updateRental(id, name, surface, price, description, userDetails);
     MsgResponseDTO response = new MsgResponseDTO("Rental updated !");
     return ResponseEntity.ok(response);
+  }
+
+  @Value("${app.upload.base-url}")
+  private String baseUrl;
+
+  @Value("${server.servlet.contextPath:/api}")
+  private String contextPath;
+
+  private RentalDTO convertToDTO(Rental rental) {
+    RentalDTO dto = new RentalDTO();
+    convertCommonParamsToDTO(rental, dto);
+    dto.setPicturesFromString(baseUrl, rental.getPicture());
+    return dto;
+  }
+
+  private RentalListDTO convertToListDTO(Rental rental) {
+    RentalListDTO dto = new RentalListDTO();
+    convertCommonParamsToDTO(rental, dto);
+    dto.setPicture(baseUrl, rental.getPicture());
+    return dto;
+  }
+
+  private RentalDTO convertCommonParamsToDTO(Rental rental, RentalDTO dto) {
+    dto.setId(rental.getId());
+    dto.setName(rental.getName());
+    dto.setSurface(rental.getSurface());
+    dto.setPrice(rental.getPrice());
+    dto.setDescription(rental.getDescription());
+    dto.setOwnerId(rental.getOwnerId());
+    dto.setCreatedAt(rental.getCreatedAt());
+    dto.setUpdatedAt(rental.getUpdatedAt());
+    return dto;
   }
 }
