+++ src/main/java/com/chatop/rental/services/UserService.java
@@ -1,6 +1,5 @@
 package com.chatop.rental.services;
 
-import com.chatop.rental.dto.UserDTO;
 import com.chatop.rental.entities.User;
 import com.chatop.rental.repositories.UserRepository;
 import org.springframework.http.HttpStatus;
@@ -15,19 +14,9 @@ public class UserService {
     this.userRepository = userRepository;
   }
 
-  public UserDTO getUserById(Long id) {
+  public User getUserById(Long id) {
     User user = userRepository.findById(id)
         .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
-    return convertToDTO(user);
-  }
-
-  public UserDTO convertToDTO(User user) {
-    UserDTO dto = new UserDTO();
-    dto.setId(user.getId());
-    dto.setName(user.getName());
-    dto.setEmail(user.getEmail());
-    dto.setCreatedAt(user.getCreatedAt());
-    dto.setUpdatedAt(user.getUpdatedAt());
-    return dto;
+    return user;
   }
 }
